Prevent tasks from being deleted after they are assigned

+++ app/graphql/mutations/delete_task.rb
@@ -6,18 +6,24 @@ class Mutations::DeleteTask < Mutations::BaseMutation
 
   def authorized?(**args)
     task = Task.find_by_uid!(args[:task])
-    policy = ApplicationPolicy.new(context[:current_user], task.application)
+    policy = TaskPolicy.new(context[:current_user], task)
     return true if policy.is_specialist_or_client
     return false, { errors: [{ code: "not_authorized" }] }
   end
 
   def resolve(**args)
     task = Task.find_by_uid!(args[:task])
+    policy = TaskPolicy.new(context[:current_user], task)
+    raise Service::Error.new("tasks.cantDeleteAssigned") unless policy.delete
+    
     task.destroy
     task.remove_from_airtable
 
     {
       task: task
     }
+
+  rescue Service::Error => e
+    { errors: [e] }
   end
 end
